refactor : QR 인증 Request Data 구성 변경

+++ app/src/main/java/com/minionz/qrna/data/QrCertificationRequestData.kt
@@ -1,6 +1,6 @@
 package com.minionz.qrna.data
 
 data class QrCertificationRequestData(
-    val userId : Int,
-    val tableId : Int
+    val userId : Long,
+    val tableId : Long
 )
